MDTable

A package for reading and writing MD tables.

Index

MDTable.readMDTable

MDTable.writeMDTable

Documentation

MDTable.readMDTable — Method

readMDTable(file::String; header::Bool=true)::DataFrame

Reads into a DataFrame any MD table file.

It's assumed the file will have a header, but can be specified with header=false if there isn't one.

MDTable.writeMDTable — Method

writeMDTable(fileName::String, df)

Creates or overwrites a file named as fileName writing in it, under MD table format, the content of any type that implements the Tables.jl interface.